server/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
aniso8601=9.0.1=pyhd3eb1b0_0
ca-certificates=2021.7.5=haa95532_1
certifi=2021.5.30=py39haa95532_0
cffi=1.14.6=pypi_0
click=8.0.1=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
cryptography=3.4.8=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
flask=1.1.2=pyhd3eb1b0_0
flask-httpauth=4.4.0=pypi_0
flask-restful=0.3.8=pyhd3eb1b0_0
flask_cors=3.0.10=pyhd3eb1b0_0
importlib-metadata=4.8.1=py39haa95532_0
itsdangerous=2.0.1=pyhd3eb1b0_0
jinja2=3.0.1=pyhd3eb1b0_0
markupsafe=2.0.1=py39h2bbff1b_0
openssl=1.1.1l=h2bbff1b_0
pip=21.2.4=py38haa95532_0
pycparser=2.20=pypi_0
pyjwt=2.1.0=py39haa95532_0
pymongo=3.12.0=py39hd77b12b_0
python=3.9.7=h6244533_1
python-dateutil=2.8.2=pyhd3eb1b0_0
pytz=2021.1=pyhd3eb1b0_0
setuptools=58.0.4=py39haa95532_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=h2bbff1b_0
tzdata=2021a=h5d7bf9c_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_1
wincertstore=0.2=py39haa95532_2
zipp=3.5.0=pyhd3eb1b0_0
